Music Player
The code that is provided embodies a basic Java Swing application that emulates the features of a straightforward music player. The user can play, advance to the next song, or go back to the previous song by navigating through a list of song titles. It's crucial to remember that this implementation only allows the song title to appear in the console; it does not allow MP3 files to be played. The System.out.println statement in the playSong method simulates this functionality by printing the title of the currently playing song to the console.